+++ examples/kitchen-sink/layout.js
@@ -8,9 +8,11 @@ export function layout(body) {
       t.title('Kitchen Sink'),
       t.link({ rel: 'icon', href: 'data:image/png;base64,iVBORw0KGgo=' }),
       t.link({ rel: 'stylesheet', href: '/styles.css' }),
-      t.script({ src: 'dist/kensington-devtools.js' }),
       t.script({ type: 'importmap' }, JSON.stringify({
-        imports: { kensington: '/dist/kensington.js' },
+        imports: {
+          kensington: '/esm/index.js',
+          'kensington/devtools': '/esm/devtools.js',
+        },
       })),
       t.script({ type: 'module', src: '/client.js' }),
     ]),